Refactor navigation helpers
